backend/requirements.txt
# Django and REST Framework
Django>=5.0
djangorestframework>=3.15.0
djangorestframework-simplejwt>=5.3.0

# Database
dj-database-url>=2.1.0
psycopg[c]; sys_platform == "linux"
psycopg[binary]; sys_platform != "linux"

# CORS
django-cors-headers>=4.3.0

# Environment variables
django-environ>=0.11.0

# Static files
whitenoise>=6.6.0

# Production server
gunicorn>=21.0.0

# Utilities
sqlparse>=0.5.0

